requirements.txt
streamlit
pdfminer.six
langchain
langchain-community
langchain-huggingface
langchain>=0.2.0
sentence-transformers
transformers
huggingface-hub
faiss-cpu
python-dotenv
torch
accelerate